requirements.txt
streamlit>=1.32.0
requests
beautifulsoup4
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
gtts
fpdf
Pillow
